backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
sentence-transformers
faiss-cpu
spacy
pdfplumber
python-docx
scikit-learn
numpy
pydantic
python-dotenv
pytesseract
pillow
transformers
torch
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
